// apps/web/src/shared/api/generated/model/booksControllerDedicationsParams.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */
import type { BooksControllerDedicationsFilter } from "./booksControllerDedicationsFilter";
import type { BooksControllerDedicationsSort } from "./booksControllerDedicationsSort";

export type BooksControllerDedicationsParams = {
  filter?: BooksControllerDedicationsFilter;
  /**
   * @minLength 1
   * @maxLength 64
   */
  genre?: string;
  /**
   * @minimum 1
   * @maximum 21474836
   */
  pageNumber?: number;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: number;
  /**
   * @maxLength 200
   */
  q?: string;
  sort?: BooksControllerDedicationsSort;
};